SSE bug is fixed in newlib
* Type casting bug is fixed

+++ lib/newlib-2.0.0/newlib/libc/string/memccpy.c
@@ -15,8 +15,8 @@ void* memccpy(void *dst, void *src, int c, size_t len) {
 	int aligned_a = 0, aligned_b = 0;
 	int i = 0;
 
-	aligned_a = ((unsigned long)a & (sizeof(__m128i) - 1));
-	aligned_b = ((unsigned long)b & (sizeof(__m128i) - 1));
+	aligned_a = ((uintptr_t)a & (sizeof(__m128i) - 1));
+	aligned_b = ((uintptr_t)b & (sizeof(__m128i) - 1));
 
 	/* Not aligned */
 	if(aligned_a != aligned_b) {
@@ -37,7 +37,7 @@ void* memccpy(void *dst, void *src, int c, size_t len) {
 
 	/* aligned */
 	if(aligned_a) {
-		while(len && ((unsigned long) &a[i] & ( sizeof(__m128i)-1))) {
+		while(len && ((uintptr_t) &a[i] & ( sizeof(__m128i)-1))) {
 			if(b[i] == endchar) {
 				a[i] = b[i];
 				return a + i;
