+++ lib/newlib-2.0.0/newlib/libc/string/memchr.c
@@ -13,10 +13,10 @@ void* memchr(void *dst, int c, size_t len) {
 	int aligned_a = 0;
 	int i = 0;
 
-	aligned_a = ((unsigned long)a & (sizeof(__m128i) - 1));
+	aligned_a = ((uintptr_t)a & (sizeof(__m128i) - 1));
 	/* aligned */
 	if(aligned_a) {
-		while(len && ((unsigned long) &a[i] & (sizeof(__m128i) - 1))) {
+		while(len && ((uintptr_t) &a[i] & (sizeof(__m128i) - 1))) {
 			if(a[i] == (char)c) {
 				return a + i;
 			}
