+++ lib/newlib-2.0.0/newlib/libc/string/memcmp.c
@@ -13,8 +13,8 @@ int memcmp(void *dst, void *src, size_t len) {
 	int aligned_a = 0, aligned_b = 0;
 	int i = 0;
 
-	aligned_a = ((unsigned long)a & (sizeof(__m128i) - 1));
-	aligned_b = ((unsigned long)b & (sizeof(__m128i) - 1));
+	aligned_a = ((uintptr_t)a & (sizeof(__m128i) - 1));
+	aligned_b = ((uintptr_t)b & (sizeof(__m128i) - 1));
 
 	/* Not aligned */
 	if(aligned_a != aligned_b) {
@@ -31,7 +31,7 @@ int memcmp(void *dst, void *src, size_t len) {
 
 	/* aligned */
 	if(aligned_a) {
-		while(len && ((unsigned long) &a[i] & (sizeof(__m128i) - 1))) {
+		while(len && ((uintptr_t) &a[i] & (sizeof(__m128i) - 1))) {
 			if(a[i] != b[i]) {
 				return b[i] - a[i];
 			}
@@ -64,7 +64,7 @@ int memcmp(void *dst, void *src, size_t len) {
 	}
 
 	while(len >= 4) {
-		if(*(long*)(&a[i]) != *(long*)(&b[i])) {
+		if(*(uint32_t*)(&a[i]) != *(uint32_t*)(&b[i])) {
 			break;
 		}
